+++ src/components/pages/Admin/components/RegistrationRequests/CompanyDetailOverlay/index.tsx
@@ -44,10 +44,11 @@ import {
 } from 'features/admin/applicationRequestApiSlice'
 import { download } from 'utils/downloadUtils'
 import CheckListFullButtons from '../components/CheckList/CheckListFullButtons'
-import { getTitle } from './CompanyDetailsHelper'
+import { getIntro, getTitle } from './CompanyDetailsHelper'
 import { useFetchNewDocumentByIdMutation } from 'features/appManagement/apiSlice'
 import { KeyValueView } from 'components/shared/basic/KeyValueView'
 import { type UniqueIdType } from 'features/admin/registration/types'
+import { StatusProgress } from '../registrationTableColumns'
 
 interface CompanyDetailOverlayProps {
   openDialog?: boolean
@@ -269,7 +270,8 @@ const CompanyDetailOverlay = ({
       >
         <DialogHeader
           {...{
-            title: getTitle(activeTab, checklistData ?? [], t),
+            title: getTitle(activeTab, t, selectedRequest),
+            intro: getIntro(activeTab, selectedCompany, t),
             closeWithIcon: true,
             onCloseWithIcon: handleOverlayClose,
           }}
@@ -277,7 +279,8 @@ const CompanyDetailOverlay = ({
           <>
             <Box
               sx={{
-                marginLeft: '60px',
+                width: '80%',
+                margin: '0 auto',
               }}
             >
               <Tabs value={activeTab} onChange={handleChange}>
@@ -285,8 +288,10 @@ const CompanyDetailOverlay = ({
                   icon={getStepIcon('1')}
                   iconPosition="start"
                   sx={{
-                    fontSize: '18px',
+                    fontSize: '14px',
                     width: '50%',
+                    textTransform: 'capitalize',
+                    color: '#111 !important',
                     '&.Mui-selected': {
                       borderBottom: '3px solid #0f71cb',
                     },
@@ -299,7 +304,10 @@ const CompanyDetailOverlay = ({
                   icon={getStepIcon('2')}
                   iconPosition="start"
                   sx={{
-                    fontSize: '18px',
+                    fontSize: '14px',
+                    width: '50%',
+                    textTransform: 'capitalize',
+                    color: '#111 !important',
                     '&.Mui-selected': {
                       borderBottom: '3px solid #0f71cb',
                     },
@@ -340,7 +348,7 @@ const CompanyDetailOverlay = ({
             }}
           >
             <TabPanel value={activeTab} index={0}>
-              <Box sx={{ width: '100%', marginBottom: '50px' }}>
+              <Box sx={{ width: '100%', margin: '50px 0' }}>
                 <KeyValueView
                   cols={2.3}
                   title={t(
@@ -364,6 +372,21 @@ const CompanyDetailOverlay = ({
             </TabPanel>
             <TabPanel value={activeTab} index={1}>
               <Box sx={{ width: '100%', height }}>
+                {selectedRequest && (
+                  <div
+                    style={{
+                      display: 'flex',
+                      justifyContent: 'center',
+                      margin: '30px',
+                    }}
+                  >
+                    <StatusProgress
+                      application={selectedRequest}
+                      trans={t}
+                      type={false}
+                    />
+                  </div>
+                )}
                 <CheckListFullButtons
                   progressButtons={checklistData}
                   selectedRequestId={selectedRequestId}